+++ exporters/variables-scss/src/content/generator.ts
@@ -1,19 +1,9 @@
-import { TokenGroup, Token, TokenType, DimensionToken } from '@supernovaio/sdk-exporters';
-import { CSSHelper, NamingHelper, StringCase } from '@supernovaio/export-helpers';
+import { TokenGroup, Token, TokenType } from '@supernovaio/sdk-exporters';
 import { exportConfiguration } from '../index';
+import { generateCssObject } from './generators/cssGenerator';
+import { dimensionTokenToCSS } from './generators/dimension';
 
 type TokenHandler = (token: Token, mappedTokens: Map<string, Token>, tokenGroups: TokenGroup[]) => string | null;
-
-const tokenVariableName = (token: Token, tokenGroups: Array<TokenGroup>, withoutParent: boolean = false): string => {
-  let parent;
-  if (withoutParent) {
-    parent = null;
-  } else {
-    parent = tokenGroups.find((group) => group.id === token.parentGroupId)!;
-  }
-
-  return NamingHelper.codeSafeVariableNameForToken(token, StringCase.paramCase, parent, '');
-};
 
 const addDisclaimer = (content: string): string => {
   if (exportConfiguration.generateDisclaimer) {
@@ -52,23 +42,6 @@ const tokensToCSS = (
     .join('\n');
 };
 
-export const dimensionTokenToCSS = (
-  token: Token,
-  mappedTokens: Map<string, Token>,
-  tokenGroups: Array<TokenGroup>,
-): string | null => {
-  if (token.tokenType !== TokenType.dimension) {
-    return null;
-  }
-
-  const dimensionToken = token as DimensionToken;
-  const name = tokenVariableName(dimensionToken, tokenGroups, true);
-  const value = dimensionToken.value?.measure ?? 0; // Add default value
-  const unit = CSSHelper.unitToCSS(dimensionToken.value?.unit);
-
-  return `$${name}: ${value}${unit} !default;`;
-};
-
 export const createSpacingContent = (
   tokens: Token[],
   mappedTokens: Map<string, Token>,
@@ -78,7 +51,7 @@ export const createSpacingContent = (
     (token) => token.tokenType === TokenType.dimension && token.origin?.name?.includes('Spacing'),
   );
   const spacingTokensToCSS = tokensToCSS(filteredSpacingTokens, dimensionTokenToCSS, mappedTokens, tokenGroups);
-  const spacingObject = generateDimensionObject(filteredSpacingTokens, mappedTokens, tokenGroups);
+  const spacingObject = generateCssObject(filteredSpacingTokens, mappedTokens, tokenGroups);
 
   return addDisclaimer(`${spacingTokensToCSS}\n\n${spacingObject}`);
 };
@@ -92,58 +65,7 @@ export const createRadiiContent = (
     (token) => token.tokenType === TokenType.dimension && token.origin?.name?.includes('Radius'),
   );
   const radiiTokensToCSS = tokensToCSS(filteredRadiiTokens, dimensionTokenToCSS, mappedTokens, tokenGroups);
-  const radiiObject = generateDimensionObject(filteredRadiiTokens, mappedTokens, tokenGroups);
+  const radiiObject = generateCssObject(filteredRadiiTokens, mappedTokens, tokenGroups);
 
   return addDisclaimer(`${radiiTokensToCSS}\n\n${radiiObject}`);
-};
-
-const generateObjectContent = (tokens: Array<Token>, tokenGroups: Array<TokenGroup>): string => {
-  let result = '';
-
-  tokens.forEach((token) => {
-    const dimensionToken = token as DimensionToken;
-    const name = tokenVariableName(dimensionToken, tokenGroups, true);
-    const numericPart = name.match(/\d+/)?.[0];
-    const prefix = `${token.origin?.name?.split('/')[0].toLowerCase()}-`; // Use template literal
-    const nonNumericPart = name.replace(prefix, '');
-    if (numericPart) {
-      result += `${numericPart}: $${name},\n`;
-    } else if (nonNumericPart) {
-      result += `${nonNumericPart}: $${name},\n`;
-    }
-  });
-
-  return result;
-};
-
-export const generateDimensionObject = (
-  tokens: Array<Token>,
-  mappedTokens: Map<string, Token>,
-  tokenGroups: Array<TokenGroup>,
-): string => {
-  const [firstToken] = tokens;
-  let objectName = '';
-  if (firstToken && firstToken.origin?.name) {
-    objectName = plural(firstToken.origin.name.split('/')[0].toLowerCase());
-  }
-
-  const objectContent = generateObjectContent(tokens, tokenGroups);
-
-  return `$${objectName}: (\n${objectContent}) !default;`;
-};
-
-const plural = (name: string): string => {
-  if (name === 'radius') {
-    return 'radii';
-  }
-
-  if (name === 'spacing') {
-    return 'spaces';
-  }
-
-  if (name.slice(-1) === 's') {
-    return name;
-  }
-
-  return `${name}s`;
 };
