+++ exporters/variables-scss/src/content/helpers/stringHelper.ts
@@ -1,0 +1,15 @@
+export const toPlural = (name: string): string => {
+  if (name === 'radius') {
+    return 'radii';
+  }
+
+  if (name === 'spacing') {
+    return 'spaces';
+  }
+
+  if (name.slice(-1) === 's') {
+    return name;
+  }
+
+  return `${name}s`;
+};
